requirements.txt
pandas>=2.1.0,<3.0.0
boto3>=1.34.0,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0
python-dotenv>=1.0.0,<2.0.0
dbt-postgres==1.5.9
pytest>=7.4.0,<9.0.0
